+++ mail/mail-uniffi/CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [unreleased] - 2025-01-XX
 
+### Added
+
+  - [ET-1633] Add async live query callback, use it for new `watch_accounts_async` and `watch_sessions_async` methods
+
 ## [0.36.0] - 2025-01-23
 
 ### Changed